Cleanup

+++ frigate/object_detection.py
@@ -10,9 +10,8 @@ from abc import ABC, abstractmethod
 import numpy as np
 from setproctitle import setproctitle
 
-from frigate.config import InputTensorEnum
 from frigate.detectors import create_detector
-
+from frigate.detectors.detector_config import InputTensorEnum
 from frigate.util import EventsPerSecond, SharedMemoryFrameManager, listen, load_labels
 
 logger = logging.getLogger(__name__)
@@ -44,7 +43,7 @@ class LocalObjectDetector(ObjectDetector):
         else:
             self.labels = load_labels(labels)
 
-        if detector_config.model.type == "object":
+        if detector_config:
             self.input_transform = tensor_transform(detector_config.model.input_tensor)
         else:
             self.input_transform = None
@@ -88,6 +87,7 @@ def run_detector(
     stop_event = mp.Event()
 
     def receiveSignal(signalNumber, frame):
+        logger.info("Signal to exit detection process...")
         stop_event.set()
 
     signal.signal(signal.SIGTERM, receiveSignal)
@@ -104,27 +104,13 @@ def run_detector(
 
     while not stop_event.is_set():
         try:
-            connection_id = detection_queue.get(timeout=5)
+            connection_id = detection_queue.get(timeout=1)
         except queue.Empty:
             continue
-        if detector_config.model.type == "audio":
-            input_frame = frame_manager.get(
-                connection_id,
-                (
-                    int(
-                        round(
-                            detector_config.model.duration
-                            * detector_config.model.sample_rate
-                        )
-                    ),
-                ),
-                dtype=np.float32,
-            )
-        else:
-            input_frame = frame_manager.get(
-                connection_id,
-                (1, detector_config.model.height, detector_config.model.width, 3),
-            )
+        input_frame = frame_manager.get(
+            connection_id,
+            (1, detector_config.model.height, detector_config.model.width, 3),
+        )
 
         if input_frame is None:
             continue
@@ -138,6 +124,8 @@ def run_detector(
         start.value = 0.0
 
         avg_speed.value = (avg_speed.value * 9 + duration) / 10
+
+    logger.info("Exited detection process...")
 
 
 class ObjectDetectProcess:
@@ -158,6 +146,9 @@ class ObjectDetectProcess:
         self.start_or_restart()
 
     def stop(self):
+        # if the process has already exited on its own, just return
+        if self.detect_process and self.detect_process.exitcode:
+            return
         self.detect_process.terminate()
         logging.info("Waiting for detection process to exit gracefully...")
         self.detect_process.join(timeout=30)
@@ -165,10 +156,11 @@ class ObjectDetectProcess:
             logging.info("Detection process didnt exit. Force killing...")
             self.detect_process.kill()
             self.detect_process.join()
+        logging.info("Detection process has exited...")
 
     def start_or_restart(self):
         self.detection_start.value = 0.0
-        if (not self.detect_process is None) and self.detect_process.is_alive():
+        if (self.detect_process is not None) and self.detect_process.is_alive():
             self.stop()
         self.detect_process = mp.Process(
             target=run_detector,
@@ -187,25 +179,19 @@ class ObjectDetectProcess:
 
 
 class RemoteObjectDetector:
-    def __init__(self, name, labels, detection_queue, event, model_config):
+    def __init__(self, name, labels, detection_queue, event, model_config, stop_event):
         self.labels = labels
         self.name = name
         self.fps = EventsPerSecond()
         self.detection_queue = detection_queue
         self.event = event
+        self.stop_event = stop_event
         self.shm = mp.shared_memory.SharedMemory(name=self.name, create=False)
-        if model_config.type == "audio":
-            self.np_shm = np.ndarray(
-                (int(round(model_config.duration * model_config.sample_rate)),),
-                dtype=np.float32,
-                buffer=self.shm.buf,
-            )
-        else:
-            self.np_shm = np.ndarray(
-                (1, model_config.height, model_config.width, 3),
-                dtype=np.uint8,
-                buffer=self.shm.buf,
-            )
+        self.np_shm = np.ndarray(
+            (1, model_config.height, model_config.width, 3),
+            dtype=np.uint8,
+            buffer=self.shm.buf,
+        )
         self.out_shm = mp.shared_memory.SharedMemory(
             name=f"out-{self.name}", create=False
         )
@@ -214,11 +200,14 @@ class RemoteObjectDetector:
     def detect(self, tensor_input, threshold=0.4):
         detections = []
 
+        if self.stop_event.is_set():
+            return detections
+
         # copy input to shared memory
         self.np_shm[:] = tensor_input[:]
         self.event.clear()
         self.detection_queue.put(self.name)
-        result = self.event.wait(timeout=10.0)
+        result = self.event.wait(timeout=5.0)
 
         # if it timed out
         if result is None:
